+++ sql/postgres/queries/withdrawal_from_processing_wallets/withdrawal_from_processing_wallets.sql
@@ -52,7 +52,7 @@ SELECT EXISTS (SELECT 1
                WHERE wfpw.request_id = $1
                  AND (
                    wfpw.transfer_id IS NULL
-                       OR t.status = 'completed'
+                       OR t.stage = 'failed'
                    ));
 
 -- name: GetByID :one